sep 15

+++ PIS2/Models/allowanceAssignmentModel.cs
@@ -15,7 +15,7 @@ namespace PIS2.Models
         [Required]
         public int employmentID { get; set; }
         public virtual employmentModel? employmentModel { get; set; }
-        public double allowanceAssignmentAmount { get; set; } = 0;
+        public decimal allowanceAssignmentAmount { get; set; } = 0;
         public DateTime? allowanceAssignmentEndDate { get; set; }
         public mainStatus allowanceStatus { get; set; } = mainStatus.Active;
         public virtual ICollection<allowanceAssignmentHistoryModel>? AllowanceAssignmentHistories { get; set; }
@@ -30,7 +30,7 @@ namespace PIS2.Models
         public virtual allowanceAssignmentModel? allowanceAssignmentModel { get; }
         public DateTime modifiedDate { get;} 
         public mainStatus allowanceAssignmentHistoryStatus { get; }
-        public double allowanceAssignmentAmount { get;}
+        public decimal allowanceAssignmentAmount { get;}
         public DateTime? allowanceAssignmentEndDate { get;  }
         public string modifiedBy { get; }
