+++ PIS2/Models/overtimeModel.cs
@@ -9,7 +9,7 @@ namespace PIS2.Models
         public string overtimeName { get; set; }
         public string?  overtimeDescription { get; set; }
         public mainStatus overtimeStatus { get; set; }
-        public double overtimeRate { get; set; }
+        public decimal overtimeRate { get; set; }
         public virtual ICollection<overtimeRecordModel>? OvertimeRecords { get; set; }
         public string modifiedBy { get; set; }
         public overtimeModel() { }
