+++ PIS2/Models/overtimeRecordModel.cs
@@ -18,15 +18,15 @@ namespace PIS2.Models
         public TimeSpan overtimeRecordStartTime { get; set; }
         public TimeSpan overtimeRecordEndTime { get; set; }
         public overtimeStatus overtimeRecordStatus { get; set; } = overtimeStatus.Hold;
-        public double overtimeRecordEmploymentRate { get; set; } = 0;
-        public double overtimeRate { get; set; } = 0;
+        public decimal overtimeRecordEmploymentRate { get; set; } = 0;
+        public decimal overtimeRate { get; set; } = 0;
         public virtual ICollection<overtimeHistoryModel>? OvertimeHistories { get; set; }
         public int? departmentID { get; set; }
         public virtual departmentModel? departmentModel {get; set;}
         public string modifiedBy { get; set; }
         public int? oldBatchNbr { get; set; }
         public TimeSpan TimeElapsed { get { return overtimeRecordEndTime - overtimeRecordStartTime; } }
-        public double GetOtCost { get { return TimeElapsed.TotalHours*overtimeRate*overtimeRecordEmploymentRate; } }  
+        public decimal GetOtCost { get { return (decimal)TimeElapsed.TotalHours*overtimeRate*overtimeRecordEmploymentRate; } }  
         public overtimeRecordModel() { }
